requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.4
bcrypt==4.0.1